clear compiler warnings

+++ dev/repitch/repitch.c
@@ -778,9 +778,9 @@ int is_peak_at(double frq,int window_offset,float minamp,dataptr dz)
 {
     float *thisbuf;
     int cc, vc, searchtop, searchbot;
-    if(window_offset) {                                                             /* BAKTRAK ALONG BIGBUF, IF NESS */
+    if(window_offset) {                                              /* BAKTRAK ALONG BIGBUF, IF NESS */
         thisbuf = dz->flbufptr[0] - (window_offset * dz->wanted);
-         if((int)thisbuf < 0 || thisbuf < dz->bigfbuf || thisbuf >= dz->flbufptr[1])
+         if((size_t) thisbuf < 0 || thisbuf < dz->bigfbuf || thisbuf >= dz->flbufptr[1])
              return(FALSE);
     } else
         thisbuf = dz->flbufptr[0];
